Extract migration logic into separate file for better organization
- Move runMigrations function to src/db/migrate.ts
- Reduce index.ts from 83 to 41 lines
- Better separation of concerns: connection setup vs migrations
- Makes codebase more maintainable and testable
